requirements.txt
Django==2.0.6
django-brfied==0.4.5
djangorestframework==3.8.2

django-extensions==2.0.7

#ipdb==0.11
